+++ sha1.c
@@ -76,14 +76,14 @@ sha1_init(struct sha1_ctx *ctx)
 
 void
 sha1_update(struct sha1_ctx *ctx,
-	    unsigned length, const uint8_t *data)
+	    size_t length, const uint8_t *data)
 {
   MD_UPDATE (ctx, length, data, COMPRESS, MD_INCR(ctx));
 }
 	  
 void
 sha1_digest(struct sha1_ctx *ctx,
-	    unsigned length,
+	    size_t length,
 	    uint8_t *digest)
 {
   uint32_t high, low;
