+++ umac96.c
@@ -48,7 +48,7 @@ umac96_set_key (struct umac96_ctx *ctx, const uint8_t *key)
 
 void
 umac96_set_nonce (struct umac96_ctx *ctx,
-		  unsigned nonce_length, const uint8_t *nonce)
+		  size_t nonce_length, const uint8_t *nonce)
 {
   assert (nonce_length > 0);
   assert (nonce_length <= AES_BLOCK_SIZE);
@@ -70,7 +70,7 @@ umac96_set_nonce (struct umac96_ctx *ctx,
 
 void
 umac96_update (struct umac96_ctx *ctx,
-	       unsigned length, const uint8_t *data)
+	       size_t length, const uint8_t *data)
 {
   MD_UPDATE (ctx, length, data, UMAC96_BLOCK, (void)0);
 }
@@ -78,7 +78,7 @@ umac96_update (struct umac96_ctx *ctx,
 
 void
 umac96_digest (struct umac96_ctx *ctx,
-	       unsigned length, uint8_t *digest)
+	       size_t length, uint8_t *digest)
 {
   uint32_t tag[4];
   unsigned i;
